Add files via upload
Chocolate, a universal delight that transcends borders and cultures. But behind every delectable bite lies a dynamic world of sales, revenue streams, and customer preferences. To understand the key drivers of success in this chocolate retail shop, we embarked on an exploratory data analysis (EDA) using MySQL, uncovering trends that influence profitability. Through meticulous data cleaning and transformation, we cleaned the dataset to extract compelling insights and recommendations to boost future performance.